recipe tests

+++ controllers/recipes.js
@@ -44,7 +44,7 @@ async function getRecipesByUser(req, res, next) {
 
 async function updateRecipe(req, res, next) { //tested without secureRoute enabled on router -> need to retest user permissions
   const id = req.params.id
-  console.log('the id inside update is: ' + id)
+  // console.log('the id inside update is: ' + id)
   const currentUser = req.currentUser
   const body = req.body
 
@@ -55,9 +55,9 @@ async function updateRecipe(req, res, next) { //tested without secureRoute enabl
       return res.send({ message: 'No recipe found' })
     }
     // check the user is the one who created the recipe || my main man masterchef
-    // if (!currentUser.isAdmin && !recipeToUpdate.user.equals(currentUser._id)){
-    //   return res.status(401).send({ message: 'Unauthorized' })
-    // }
+    if (!currentUser.isAdmin && !recipeToUpdate.user.equals(currentUser._id)){
+      return res.status(401).send({ message: 'Unauthorized' })
+    }
     recipeToUpdate.set(body)
     recipeToUpdate.save()
     res.send(recipeToUpdate)
@@ -82,9 +82,9 @@ async function deleteRecipe(req, res, next) { //tested without secureRoute enabl
       return res.send({ message: 'No recipe found' })
     }
     // check the user is the one who created the recipe || my main man masterchef
-    // if (!currentUser.isAdmin && !recipeToDelete.user.equals(currentUser._id)){
-    //   return res.status(401).send({ message: 'Unauthorized' })
-    // }
+    if (!currentUser.isAdmin && !recipeToDelete.user.equals(currentUser._id)){
+      return res.status(401).send({ message: 'Unauthorized' })
+    }
     await recipeToDelete.deleteOne()
 
     res.send(recipeToDelete)
@@ -99,7 +99,7 @@ async function deleteRecipe(req, res, next) { //tested without secureRoute enabl
 export default {
   getRecipes,
   updateRecipe,
-  deleteRecipe
+  deleteRecipe,
   getSingleRecipe,
   makeRecipe,
   getRecipesByUser
